requirements.txt
Flask==2.3.3
joblib==1.3.0
numpy==1.24.3
Pillow==9.5.0
tensorflow==2.13.0
scikit-learn==1.3.0